update segment ot axe distance of root.tracking
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/vplants/vplants/branches/rhizoscan@15495 ab253cce-29fb-0310-bb2f-979600cdbdeb

+++ src/rhizoscan/root/tracking.py
@@ -82,7 +82,7 @@ def segment_to_axe_distance(graph,tree):
     ta_end = _np.cumsum(ta_end)
     
     # compute the minimlum segment distance per axe
-    d_na = _np.empty(d_ns.shape[0],len(ta_end))
+    d_na = _np.empty((d_ns.shape[0],len(ta_end)),dtype=d_ns.dtype)
     start = 0
     for i,end in enumerate(ta_end):
         d_na[i] = d_ns[:,start:end].min(axis=0)
